Update the onboarding button

+++ BatFiKit/Sources/DefaultsKeys/Defaults.swift
@@ -10,7 +10,7 @@ import Foundation
 
 extension Defaults.Keys {
     public static let launchAtLogin = Key<Bool>("launchAtLogin", default: true)
-    public static let chargeLimit = Key<Double>("chargeLimit", default: 80)
+    public static let chargeLimit = Key<Int>("chargeLimit", default: 80)
     public static let manageCharging = Key<Bool>("manageCharging", default: true)
     public static let temperatureSwitch = Key<Bool>("temperatureSwitch", default: true)
     public static let disableCharging = Key<Bool>("disableCharging", default: false)
